pyproject.toml
[tool.poetry]
name = "llamaso"
version = "0.0.1"
description = "A Python package for LlamaCPP for Structured Output."
readme = "README.md"
authors = ["Kanishk Navale <navalekanishk@gmail.com>"]
license = "MIT"
packages = [{ include = "llamaso" }]

[tool.poetry.dependencies]
python = "^3.10"
isort = ">=5.13.2"
llama-cpp-python = ">=0.3.1"
pydantic = ">=2.9.2"
ruff = ">=0.7.2"
huggingface-hub = "^0.26.2"
json-repair = "^0.30.0"

[tool.isort]
atomic = true
force_grid_wrap = 0
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
multi_line_output = 3
not_skip = "__init__.py"
use_parentheses = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
